% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4_coefMerMod_new.R
\name{coef.merMod_new}
\alias{coef.merMod_new}
\title{slightly adapted version of lme4:::coef.merMod}
\usage{
\method{coef}{merMod_new}(object, ...)
}
\description{
slightly adapted version of lme4:::coef.merMod
}
